// src/utils/wordTools/types.ts
export interface WordToolDefinition {
  name: string
  description: string
  inputSchema: {
    type: 'object'
    properties: Record<string, any>
    required?: string[]
  }
  execute: (args: any) => Promise<any>
}

export type WordToolName =
  | 'getSelectedText'
  | 'getDocumentContent'
  | 'insertText'
  | 'replaceSelectedText'
  | 'appendText'
  | 'insertParagraph'
  | 'formatText'
  | 'searchAndReplace'
  | 'getDocumentProperties'
  | 'insertTable'
  | 'insertList'
  | 'deleteText'
  | 'clearFormatting'
  | 'setFontName'
  | 'insertPageBreak'
  | 'getRangeInfo'
  | 'selectText'
  | 'insertImage'
  | 'getTableInfo'
  | 'insertBookmark'
  | 'goToBookmark'
  | 'insertContentControl'
  | 'findText'
  | 'bulkFindReplace'
  | 'applyStyle'
  | 'createSection'
  | 'formatTable'
  | 'getDocumentStructure'
  | 'setPageMargins'
  | 'setPageOrientation'
  | 'setParagraphAlignment'
  | 'setLineSpacing'
  | 'setHeaderText'
  | 'setFooterText'
  | 'insertComment'
  | 'listComments'
  | 'deleteComment'
  | 'setTrackChanges'
  | 'insertHyperlink'
  | 'insertFootnote'
  | 'insertEndnote'
  | 'insertTableOfContents'
  | 'listStyles'
  | 'insertPageNumber'
  | 'setParagraphSpacing'
  | 'setParagraphIndentation'
  | 'setParagraphBorders'
  | 'setParagraphShading'
  | 'setFontAdvanced'
  | 'insertSymbol'
  | 'setPageColumnCount'
  | 'acceptTrackedChanges'
  | 'rejectTrackedChanges'
  | 'insertTextBox'
  | 'saveDocument'
  | 'printDocument'
  | 'exportAsPdf'
  | 'newDocument'
  | 'changeCase'
  | 'insertShape'
  | 'insertChart'
  | 'insertSmartArt'
  | 'setPageColor'
  | 'insertWatermark'
  | 'setPaperSize'
  | 'setLineNumbers'
  | 'insertCaption'
  | 'insertIndex'
  | 'insertCitation'
  | 'insertTableOfAuthorities'
  | 'compareDocuments'
  | 'protectDocument'
  | 'setZoom'
  | 'setHyphenation'
  | 'arrangeObject'
  | 'translateSelection'
  | 'insertAnnotation'
  | 'insertCheckboxContentControl'
  | 'updateCheckboxContentControl'
  | 'manageField'
  | 'mergeTableCells'
  | 'splitTableCells'